Use default branch instead of master #14
Remove hardcoded 'master' branch from `git push` calls. It allows
flexibility of working with any default branch naming such as `main`.

Remove branch name when pushing tag. A tag and a branch don't have a
direct relationship. Pushing a tag does not have anything to do with
any branch. The tag points to a commit. It does not care which branch
the commit is reachable from. The commit may be on master and/or on
another branch but it's none of the tag's business.

Remove hardcored master when pushing changes. `git push` without a
specified branch will push changes on the local branch to matching
remote branch anyway.

Add related tests and extend test logic to support the new tests.